timetracker/backend/requirements.txt
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
openai>=1.0,<2.0
python-dotenv>=1.0,<2.0
requests>=2.31,<3.0
numpy>=1.26,<3.0
beautifulsoup4>=4.12,<5.0
